+++ src/factories/IManagerFactory.ts
@@ -1,5 +1,5 @@
 import { IManager } from '../entities/IManager';
-import Manager from '../entities/Manager';
+import Manager from '../entities/implementations/Manager';
 
 export interface IManagerFactory {
   make(manager: IManager): Manager
