+++ src/factories/implementations/CreateManagerControllerFactory.ts
@@ -1,11 +1,11 @@
-import CreateManagerController from '../controllers/CreateManagerController';
-import prismaModel from '../database/prisma';
-import TokenProvider from '../providers/implementations/TokenProviderAdapter';
+import CreateManagerController from '../../controllers/CreateManagerController';
+import prismaModel from '../../database/prisma';
+import TokenProvider from '../../providers/implementations/TokenProviderAdapter';
 import {
   createManagerSchema,
-} from '../providers/implementations/zodValidator/schemas/CreateManager';
-import ManagerRepository from '../repositories/implementations/ManagerRepository';
-import CreateManagerService from '../services/CreateManagerService';
+} from '../../providers/implementations/zodValidator/schemas/CreateManager';
+import ManagerRepository from '../../repositories/implementations/ManagerRepository';
+import CreateManagerService from '../../services/CreateManagerService';
 import ManagerFactory from './ManagerFactory';
 
 export default class CreateManagerControllerFactory {
