+++ src/factories/implementations/ManagerFactory.ts
@@ -1,7 +1,7 @@
 /* eslint-disable class-methods-use-this */
-import { IManager } from '../entities/IManager';
-import Manager from '../entities/Manager';
-import { IManagerFactory } from './IManagerFactory';
+import { IManager } from '../../entities/IManager';
+import Manager from '../../entities/implementations/Manager';
+import { IManagerFactory } from '../IManagerFactory';
 
 export default class ManagerFactory implements IManagerFactory {
   make(manager: IManager): Manager {
