+++ src/routes/ManagerRoutes.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import CreateManagerControllerFactory from '../factories/CreateManagerControllerFactory';
+import CreateManagerControllerFactory from '../factories/implementations/CreateManagerControllerFactory';
 
 const createManagerController = CreateManagerControllerFactory.make();
